make wasi-common self-contained, cut dep from wasmtime-wasi
