+++ crates/wasi-common/tests/all/sync.rs
@@ -1,6 +1,7 @@
 use super::*;
 use test_programs_artifacts::*;
-use wasmtime_wasi::sync::{add_to_linker, WasiCtxBuilder};
+use wasi_cap_std_sync::WasiCtxBuilder;
+use wasi_common::sync::add_to_linker;
 
 foreach_preview1!(assert_test_exists);
