+++ crates/wasi-threads/Cargo.toml
@@ -18,8 +18,5 @@ workspace = true
 anyhow = { workspace = true }
 log = { workspace = true }
 rand = "0.8"
-wasi-common = { workspace = true }
+wasi-common = { workspace = true, features = ["exit"] }
 wasmtime = { workspace = true }
-wasmtime-wasi = { workspace = true, default-features = true, features = [
-    "exit",
-] }
